Resolves #3: renaming files
